# optimize(tea): apply fast speed from MiraiGo