--- a/packets/wtlogin/sso.go
+++ b/packets/wtlogin/sso.go
@@ -7,11 +7,10 @@ import (
 	"strings"
 
 	"github.com/LagrangeDev/LagrangeGo/utils"
+	"github.com/LagrangeDev/LagrangeGo/utils/crypto"
 	"github.com/LagrangeDev/LagrangeGo/utils/crypto/ecdh"
 
 	binary2 "github.com/LagrangeDev/LagrangeGo/utils/binary"
-
-	qqtea "github.com/LagrangeDev/LagrangeGo/utils/crypto"
 )
 
 type SSOPacket struct {
@@ -66,11 +65,11 @@ func ParseSSOHeader(raw, d2Key []byte) (*SSOHeader, error) {
 	} else if flag == 1 { // enc with d2key
 		temp := make([]byte, buf.Len())
 		_ = binary.Read(buf, binary.BigEndian, temp)
-		dec = qqtea.QQTeaDecrypt(temp, d2Key)
+		dec = crypto.NewTeaCipher(d2Key).Decrypt(temp)
 	} else if flag == 2 { // enc with \x00*16
 		temp := make([]byte, buf.Len())
 		_ = binary.Read(buf, binary.BigEndian, temp)
-		dec = qqtea.QQTeaDecrypt(temp, make([]byte, 16))
+		dec = crypto.NewTeaCipher(nil).Decrypt(temp)
 	} else {
 		return nil, fmt.Errorf("invalid encrypt flag: %d", flag)
 	}
@@ -137,7 +136,7 @@ func ParseOicqBody(buf []byte) ([]byte, error) {
 
 	body := buf[16 : len(buf)-1]
 	if encType == 0 {
-		return qqtea.QQTeaDecrypt(body, ecdh.ECDH["secp192k1"].GetShareKey()), nil
+		return crypto.NewTeaCipher(ecdh.ECDH["secp192k1"].GetShareKey()).Decrypt(body), nil
 	} else {
 		return nil, fmt.Errorf("unknown encrypt type: %d", encType)
 	}