--- a/utils/crypto/tea.go
+++ b/utils/crypto/tea.go
@@ -1,210 +1,145 @@
 package crypto
 
+// from https://github.com/Mrs4s/MiraiGo/blob/master/binary/tea.go
+
 import (
-	"bytes"
 	"encoding/binary"
+	_ "unsafe" // required by go:linkname
 )
 
-// 使用utils里面的reader会循环引用
+type TEA [4]uint32
 
-type Reader struct {
-	buffer []byte
-	pos    int
-}
+// Encrypt tea 加密
+// http://bbs.chinaunix.net/thread-583468-1-1.html
+// 感谢xichen大佬对TEA的解释
+func (t TEA) Encrypt(src []byte) (dst []byte) {
+	lens := len(src)
+	fill := 10 - (lens+1)%8
+	dst = make([]byte, fill+lens+7)
+	dst[0] = byte(fill-3) | 0xF8 // 存储pad长度
+	copy(dst[fill:], src)
 
-func NewReader(buffer []byte) *Reader {
-	return &Reader{
-		buffer: buffer,
-		pos:    0,
-	}
-}
-
-func (r *Reader) ReadU32() (v uint32) {
-	v = binary.BigEndian.Uint32(r.buffer[r.pos : r.pos+4])
-	r.pos += 4
-	return
-}
-
-const (
-	DELTA = uint64(0x9E3779B9)
-	ROUND = 16
-	OP    = uint64(0xFFFFFFFF)
-)
-
-func xor(a, b []byte) []byte {
-	var a1, a2, b1, b2 uint32
-	readera := NewReader(a)
-	readerb := NewReader(b)
-
-	a1 = readera.ReadU32()
-	a2 = readera.ReadU32()
-	b1 = readerb.ReadU32()
-	b2 = readerb.ReadU32()
-	//_ = binary.Read(readera, binary.BigEndian, &a1)
-	//_ = binary.Read(readera, binary.BigEndian, &a2)
-	//_ = binary.Read(readerb, binary.BigEndian, &b1)
-	//_ = binary.Read(readerb, binary.BigEndian, &b2)
-
-	buf := make([]byte, 0)
-	buf = binary.BigEndian.AppendUint32(buf, (a1^b1)&uint32(OP))
-	buf = binary.BigEndian.AppendUint32(buf, (a2^b2)&uint32(OP))
-
-	return buf
-}
-
-func teaCode(v, k []byte) []byte {
-	var key0, key1, key2, key3, val0, val1 uint32
-	var k0, k1, k2, k3, v0, v1 uint64
-
-	//reader1 := bytes.NewReader(k)
-	//_ = binary.Read(reader1, binary.BigEndian, &key0)
-	//_ = binary.Read(reader1, binary.BigEndian, &key1)
-	//_ = binary.Read(reader1, binary.BigEndian, &key2)
-	//_ = binary.Read(reader1, binary.BigEndian, &key3)
-
-	reader1 := NewReader(k)
-	key0 = reader1.ReadU32()
-	key1 = reader1.ReadU32()
-	key2 = reader1.ReadU32()
-	key3 = reader1.ReadU32()
-
-	k0 = uint64(key0)
-	k1 = uint64(key1)
-	k2 = uint64(key2)
-	k3 = uint64(key3)
-
-	//reader2 := bytes.NewReader(v)
-	//_ = binary.Read(reader2, binary.BigEndian, &val0)
-	//_ = binary.Read(reader2, binary.BigEndian, &val1)
-
-	reader2 := NewReader(v)
-	val0 = reader2.ReadU32()
-	val1 = reader2.ReadU32()
-
-	v0 = uint64(val0)
-	v1 = uint64(val1)
-
-	sum := uint64(0)
-	for i := 0; i < ROUND; i++ {
-		sum += DELTA
-		v0 += ((OP & (v1 << 4)) + k0) ^ (v1 + sum) ^ ((OP & (v1 >> 5)) + k1)
-		v0 &= OP
-		v1 += ((OP & (v0 << 4)) + k2) ^ (v0 + sum) ^ ((OP & (v0 >> 5)) + k3)
-		v1 &= OP
+	var iv1, iv2, holder uint64
+	for i := 0; i < len(dst); i += 8 {
+		block := binary.BigEndian.Uint64(dst[i:])
+		holder = block ^ iv1
+		iv1 = t.encode(holder)
+		iv1 ^= iv2
+		iv2 = holder
+		binary.BigEndian.PutUint64(dst[i:], iv1)
 	}
 
-	buf := make([]byte, 0)
-	buf = binary.BigEndian.AppendUint32(buf, uint32(v0))
-	buf = binary.BigEndian.AppendUint32(buf, uint32(v1))
-
-	return buf
+	return dst
 }
 
-func teaDecipher(v, k []byte) []byte {
-	var key0, key1, key2, key3, val0, val1 uint32
-	var k0, k1, k2, k3, v0, v1 uint64
-
-	//reader1 := bytes.NewReader(k)
-	//_ = binary.Read(reader1, binary.BigEndian, &key0)
-	//_ = binary.Read(reader1, binary.BigEndian, &key1)
-	//_ = binary.Read(reader1, binary.BigEndian, &key2)
-	//_ = binary.Read(reader1, binary.BigEndian, &key3)
-
-	reader1 := NewReader(k)
-	key0 = reader1.ReadU32()
-	key1 = reader1.ReadU32()
-	key2 = reader1.ReadU32()
-	key3 = reader1.ReadU32()
-
-	k0 = uint64(key0)
-	k1 = uint64(key1)
-	k2 = uint64(key2)
-	k3 = uint64(key3)
-
-	//reader2 := bytes.NewReader(v)
-	//_ = binary.Read(reader2, binary.BigEndian, &val0)
-	//_ = binary.Read(reader2, binary.BigEndian, &val1)
-
-	reader2 := NewReader(v)
-	val0 = reader2.ReadU32()
-	val1 = reader2.ReadU32()
-
-	v0 = uint64(val0)
-	v1 = uint64(val1)
-
-	sum := DELTA << 4 & OP
-	for i := 0; i < ROUND; i++ {
-		v1 -= ((v0 << 4) + k2) ^ (v0 + sum) ^ ((v0 >> 5) + k3)
-		v1 &= OP
-		v0 -= ((v1 << 4) + k0) ^ (v1 + sum) ^ ((v1 >> 5) + k1)
-		v0 &= OP
-		sum -= DELTA
-		sum &= OP
-	}
-
-	buf := make([]byte, 0)
-	buf = binary.BigEndian.AppendUint32(buf, uint32(v0))
-	buf = binary.BigEndian.AppendUint32(buf, uint32(v1))
-
-	return buf
-}
-
-func preprocess(data []byte) []byte {
-	dataLen := len(data)
-	// 保证取余出来是正数
-	filln := ((8-(dataLen+2))%8+8)%8 + 2
-	fills := make([]byte, 0)
-	for i := 0; i < filln; i++ {
-		fills = append(fills, 220)
-	}
-	return append(append(append([]byte{byte((filln - 2) | 0xf8)}, fills...), data...), make([]byte, 7)...)
-}
-
-func encrypt(data, key []byte) []byte {
-	if len(key) < 16 {
-		panic("key length must be 16")
-	}
-
-	data = preprocess(data)
-	tr := make([]byte, 8)
-	to := make([]byte, 8)
-	result := make([]byte, 0)
-	for i := 0; i < len(data); i += 8 {
-		o := xor(data[i:i+8], tr)
-		tr = xor(teaCode(o, key), to)
-		to = o
-		result = append(result, tr...)
-	}
-	return result
-}
-
-func decrypt(text, key []byte) []byte {
-	if len(key) < 16 {
-		panic("key length must be 16")
-	}
-
-	dataLen := len(text)
-	plain := teaDecipher(text, key)
-	pos := (plain[0] & 0x07) + 2
-	ret := plain
-	precrypt := text[0:8]
-	for i := 8; i < dataLen; i += 8 {
-		x := xor(teaDecipher(xor(text[i:i+8], plain), key), precrypt)
-		plain = xor(x, precrypt)
-		precrypt = text[i : i+8]
-		ret = append(ret, x...)
-	}
-	if !bytes.Equal(ret[len(ret)-7:], make([]byte, 7)) {
+func (t TEA) Decrypt(data []byte) []byte {
+	if len(data) < 16 || len(data)%8 != 0 {
 		return nil
 	}
-
-	return ret[pos+1 : len(ret)-7]
+	dst := make([]byte, len(data))
+	var iv1, iv2, holder uint64
+	for i := 0; i < len(dst); i += 8 {
+		iv1 = binary.BigEndian.Uint64(data[i:])
+		iv2 ^= iv1
+		iv2 = t.decode(iv2)
+		binary.BigEndian.PutUint64(dst[i:], iv2^holder)
+		holder = iv1
+	}
+	return dst[dst[0]&7+3 : len(data)-7]
 }
 
-func QQTeaEncrypt(data, key []byte) []byte {
-	return encrypt(data, key)
+//go:nosplit
+func (t *TEA) encode(n uint64) uint64 {
+	v0, v1 := uint32(n>>32), uint32(n)
+	t0, t1, t2, t3 := t[0], t[1], t[2], t[3]
+
+	v0 += (v1 + 0x9e3779b9) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x9e3779b9) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x3c6ef372) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x3c6ef372) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0xdaa66d2b) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0xdaa66d2b) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x78dde6e4) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x78dde6e4) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x1715609d) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x1715609d) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0xb54cda56) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0xb54cda56) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x5384540f) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x5384540f) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0xf1bbcdc8) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0xf1bbcdc8) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x8ff34781) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x8ff34781) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x2e2ac13a) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x2e2ac13a) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0xcc623af3) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0xcc623af3) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x6a99b4ac) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x6a99b4ac) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x08d12e65) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x08d12e65) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0xa708a81e) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0xa708a81e) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0x454021d7) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0x454021d7) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 += (v1 + 0xe3779b90) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 += (v0 + 0xe3779b90) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+
+	return uint64(v0)<<32 | uint64(v1)
 }
 
-func QQTeaDecrypt(data, key []byte) []byte {
-	return decrypt(data, key)
+// 每次8字节
+//
+//go:nosplit
+func (t *TEA) decode(n uint64) uint64 {
+	v0, v1 := uint32(n>>32), uint32(n)
+	t0, t1, t2, t3 := t[0], t[1], t[2], t[3]
+
+	v1 -= (v0 + 0xe3779b90) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0xe3779b90) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x454021d7) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x454021d7) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0xa708a81e) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0xa708a81e) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x08d12e65) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x08d12e65) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x6a99b4ac) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x6a99b4ac) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0xcc623af3) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0xcc623af3) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x2e2ac13a) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x2e2ac13a) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x8ff34781) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x8ff34781) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0xf1bbcdc8) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0xf1bbcdc8) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x5384540f) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x5384540f) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0xb54cda56) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0xb54cda56) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x1715609d) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x1715609d) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x78dde6e4) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x78dde6e4) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0xdaa66d2b) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0xdaa66d2b) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x3c6ef372) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x3c6ef372) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+	v1 -= (v0 + 0x9e3779b9) ^ (v0<<4 + t2) ^ (v0>>5 + t3)
+	v0 -= (v1 + 0x9e3779b9) ^ (v1<<4 + t0) ^ (v1>>5 + t1)
+
+	return uint64(v0)<<32 | uint64(v1)
+}
+
+//go:nosplit
+func NewTeaCipher(key []byte) (t TEA) {
+	if len(key) != 16 {
+		return TEA{}
+	}
+	t[3] = binary.BigEndian.Uint32(key[12:16])
+	t[2] = binary.BigEndian.Uint32(key[8:12])
+	t[1] = binary.BigEndian.Uint32(key[4:8])
+	t[0] = binary.BigEndian.Uint32(key[:4])
+	return t
 }