应用经过优化的TEA。
```
goos: darwin
goarch: arm64
pkg: github.com/LagrangeDev/LagrangeGo/utils/crypto
            │ utils/crypto/old.txt │        utils/crypto/new.txt         │
            │        sec/op        │   sec/op     vs base                │
TEAen/16-8             195.5n ± 0%   225.7n ± 3%  +15.42% (p=0.000 n=10)
TEAen/256-8            3.734µ ± 0%   1.924µ ± 1%  -48.47% (p=0.000 n=10)
TEAen/4K-8             58.59µ ± 0%   29.05µ ± 0%  -50.42% (p=0.000 n=10)
TEAen/32K-8            468.9µ ± 1%   228.3µ ± 0%  -51.30% (p=0.000 n=10)
TEAde/16-8             452.9n ± 0%   254.7n ± 0%  -43.76% (p=0.000 n=10)
TEAde/256-8            4.046µ ± 0%   2.218µ ± 1%  -45.19% (p=0.000 n=10)
TEAde/4K-8             59.70µ ± 0%   33.56µ ± 0%  -43.77% (p=0.000 n=10)
TEAde/32K-8            474.1µ ± 0%   268.0µ ± 1%  -43.48% (p=0.000 n=10)
geomean                13.40µ        7.864µ       -41.32%

            │ utils/crypto/old.txt │          utils/crypto/new.txt          │
            │         B/s          │      B/s       vs base                 │
TEAen/16-8            78.05Mi ± 0%    67.61Mi ± 3%   -13.37% (p=0.000 n=10)
TEAen/256-8           65.39Mi ± 0%   126.89Mi ± 1%   +94.04% (p=0.000 n=10)
TEAen/4K-8            66.67Mi ± 0%   134.48Mi ± 0%  +101.72% (p=0.000 n=10)
TEAen/32K-8           66.65Mi ± 1%   136.86Mi ± 0%  +105.33% (p=0.000 n=10)
TEAde/16-8            67.39Mi ± 0%   119.82Mi ± 0%   +77.81% (p=0.000 n=10)
TEAde/256-8           64.11Mi ± 0%   116.98Mi ± 1%   +82.46% (p=0.000 n=10)
TEAde/4K-8            65.69Mi ± 0%   116.84Mi ± 0%   +77.87% (p=0.000 n=10)
TEAde/32K-8           65.95Mi ± 0%   116.67Mi ± 1%   +76.92% (p=0.000 n=10)
geomean               67.37Mi         114.8Mi        +70.40%

            │ utils/crypto/old.txt │         utils/crypto/new.txt         │
            │         B/op         │     B/op      vs base                │
TEAen/16-8              56.00 ± 0%     32.00 ± 0%  -42.86% (p=0.000 n=10)
TEAen/256-8            1496.0 ± 0%     288.0 ± 0%  -80.75% (p=0.000 n=10)
TEAen/4K-8           28.211Ki ± 0%   4.750Ki ± 0%  -83.16% (p=0.000 n=10)
TEAen/32K-8          278.71Ki ± 0%   40.00Ki ± 0%  -85.65% (p=0.000 n=10)
TEAde/16-8             152.00 ± 0%     32.00 ± 0%  -78.95% (p=0.000 n=10)
TEAde/256-8            2072.0 ± 0%     288.0 ± 0%  -86.10% (p=0.000 n=10)
TEAde/4K-8           33.523Ki ± 0%   4.750Ki ± 0%  -85.83% (p=0.000 n=10)
TEAde/32K-8          278.78Ki ± 0%   40.00Ki ± 0%  -85.65% (p=0.000 n=10)
geomean               6.037Ki        1.137Ki       -81.17%

            │ utils/crypto/old.txt │        utils/crypto/new.txt        │
            │      allocs/op       │ allocs/op   vs base                │
TEAen/16-8              6.000 ± 0%   1.000 ± 0%  -83.33% (p=0.000 n=10)
TEAen/256-8           130.000 ± 0%   1.000 ± 0%  -99.23% (p=0.000 n=10)
TEAen/4K-8           2056.000 ± 0%   1.000 ± 0%  -99.95% (p=0.000 n=10)
TEAen/32K-8         16400.000 ± 0%   1.000 ± 0%  -99.99% (p=0.000 n=10)
TEAde/16-8             15.000 ± 0%   1.000 ± 0%  -93.33% (p=0.000 n=10)
TEAde/256-8           139.000 ± 0%   1.000 ± 0%  -99.28% (p=0.000 n=10)
TEAde/4K-8           2065.000 ± 0%   1.000 ± 0%  -99.95% (p=0.000 n=10)
TEAde/32K-8         16408.000 ± 0%   1.000 ± 0%  -99.99% (p=0.000 n=10)
geomean                 455.6        1.000       -99.78%
```